some fine tuning and data type conversion issues

+++ MotorController.h
@@ -59,7 +59,7 @@ private:
     EEPROMConfig &_eepromConfig;
 
     void setTargetSpeed(double speed);
-    int readEncoder(); // Method to read the encoder position
+    uint16_t readEncoder(); // Method to read the encoder position
     double rpmToEncoderCountsPerSecond(double rpm);
     void readGUID(char *guid);
     void setDirection(String direction);
